Move all MySQL credentials to a single location in values.yaml
This move all mysql credentials to `mysql.users.*` and allows accounts
to be changed in a single location. Additionally, credentials for each
service (API, notification engine, thresh, keystone) are now stored in
secrets.

Note that thresh does not yet yet make use of its secret as the
credentials currently must be kept in a ConfigMap. This will be
resolved once the storm/thresh charts are updated to use the
refactored images.

Additionally, this removes the configurable `mysqlDatabase` parameter
as it wasn't truly configurable. Changing it from the default would
break deployment in (almost) all cases as mysql-init doesn't honor the
parameter, and more generally it isn't something that needs to be
user-configurable.