Use release way of specifying storage class
The previous zookeeper and kafka charts a combination of alpha
and beta annotations for setting the storage class. The release
way of doing that is setting the spec property storageClassName.

This breaks helm upgrade if persistent volumes are used so the
minor version is updated on those two charts and monasca which is
changed to use the new charts.