Add seperate flags for controller and tpr
We need to do this as if we are deploying the monasca chart
at the same time we are deploying other charts that rely on the
tpr then we will get into timing issues. This could happen using
the helm wrapper tool landscaper. By allowing seperate flags
we could just enable the controller and not the resource and assume
the resource is created externally before deploy