Add optional serviceAccount parameter for the cleanup hook
The cleanup hook makes it impossible to upgrade to a cluster with
RBAC as its own ServiceAccount will not exist before the cleanup job
itself is run. There are two possible solutions to this:

 - Use `helm upgrade --no-hooks` to skip running hooks for the
   initial upgrade. helm will add the necessary service account and
   RBAC rules during the upgrade, though the cleanup job will be
   skipped. Future upgrades will work as expected.
 - Allow users to specify a preexisting serviceAccount for the
   cleanup hook.

This patch adds support for the second option. Users can opt to
upgrade once with `--no-hooks` instead if desired, or create a
service account for hooks in advance to avoid edge cases.