[monasca] overrideUri for kafka/zookeeper, non-local traffic in agent
Adds support for `kafka.overrideUri` and `zookeeper.overrideUri`. If
set, these values allow an external kafka/zookeeper to be used.

Also adds support for setting the `NON_LOCAL_TRAFFIC` variable in
monasca-agent. On some systems with strange IPv6 configuration, the
forwarder is unable to bind to `localhost` properly. Enabling
non-local traffic makes the forwarder bind to 0.0.0.0 instead, which
works around the problem.

Signed-off-by: Tim Buckley <timothy.jas.buckley@hpe.com>