[monasca-agent] Set `NON_LOCAL_TRAFFIC` in forwarder
Also adds support for setting the `NON_LOCAL_TRAFFIC` variable in
monasca-agent. On some systems with strange IPv6 configuration, the
forwarder is unable to bind to `localhost` properly. Enabling
non-local traffic makes the forwarder bind to 0.0.0.0 instead, which
works around the problem. As we're running in a container and no
ports are exposed by default, this workaround should be safe to apply
automatically.

Signed-off-by: Tim Buckley <timothy.jas.buckley@hpe.com>